New bridge.py file for NapariBridge class
