+++ webmon.py
@@ -18,8 +18,7 @@ import logging
 import os
 import sys
 from pathlib import Path
-from queue import Empty, Queue
-from threading import Lock, get_ident
+from threading import Lock
 from typing import Optional
 
 import click
@@ -27,6 +26,7 @@ import requests
 from flask import Flask, render_template, session
 from flask_socketio import Namespace, SocketIO, emit
 
+from bridge import NapariBridge
 from lib.numpy_json import NumpyJSON
 from napari_client import NapariClient
 
@@ -73,56 +73,13 @@ socketio = SocketIO(app, async_mode=ASYNC_MODE, json=NumpyJSON)
 thread = None
 thread_lock = Lock()
 
-# Number of seconds between updates
-poll_interval_seconds = 0.01
-
-
-class NapariClientManager:
-    def __init__(self, client: NapariClient):
-        self.client = client
-        self.commands = Queue()
-
-    def send_command(self, command) -> None:
-        """Put into queue for background_task to send."""
-        self.commands.put(command)
-
-    def background_task(self) -> None:
-        """Send data to/from the viewer and napari."""
-        tid = get_ident()
-        LOGGER.info("Webmon: Background task thread_id=%d", tid)
-
-        while True:
-            socketio.sleep(poll_interval_seconds)
-
-            if self.client is None:
-                continue
-
-            self._send_commands()  # Send any pending commands.
-
-            # We just send the whole thing every time. We could potentially
-            # only send if the data had changed since the last time
-            # we sent it. To cut down the spamming viewer with repeated
-            # data.
-            tile_data = self.client.napari_data['tile_data']
-            socketio.emit('set_tile_data', tile_data, namespace='/test')
-
-    def _send_commands(self) -> None:
-        """Send all pending commands."""
-        while True:
-            try:
-                command = self.commands.get_nowait()
-            except Empty:
-                break  # No more commands to end.
-            LOGGER.info("Send command to napari: %s", command)
-            self.client.send_command(command)
-
 
 class WebmonHandlers(Namespace):
     """Generic handlers right now, but will be per-page soon."""
 
-    def __init__(self, namespace: str, manager: NapariClientManager):
+    def __init__(self, namespace: str, bridge: NapariBridge):
         super().__init__(namespace)
-        self.manager = manager
+        self.bridge = bridge
 
     def on_connection_test(self, message):
         LOGGER.info("connection_test: %s", message)
@@ -138,7 +95,7 @@ class WebmonHandlers(Namespace):
         emit('input_data_response', json.dumps(data))
 
     def on_gui_input(self, message):
-        self.manager.send_command(message)
+        self.bridge.send_command(message)
 
     def on_connect(self):
         LOGGER.info("on_connect")
@@ -150,7 +107,7 @@ class WebmonHandlers(Namespace):
             if thread is None:
                 LOGGER.info("Webmon: Creating background task...")
                 thread = socketio.start_background_task(
-                    target=self.manager.background_task
+                    target=self.bridge.background_task
                 )
 
 
@@ -275,9 +232,9 @@ def main(log_path: Optional[str], port: int) -> None:
     global client
     client = _create_napari_client(port)
 
-    manager = NapariClientManager(client)
+    bridge = NapariBridge(socketio, client)
 
-    socketio.on_namespace(WebmonHandlers('/test', manager))
+    socketio.on_namespace(WebmonHandlers('/test', bridge))
 
     # socketio.run does not exit until our /stop endpoint is hit.
     socketio.run(
